Move UserAssignment class to meta package
