+++ server/src/main/java/io/littlehorse/common/model/meta/UserTaskNode.java
@@ -1,11 +1,10 @@
-package io.littlehorse.common.model.meta.subnode;
+package io.littlehorse.common.model.meta;
 
 import com.google.protobuf.Message;
 import io.littlehorse.common.LHConfig;
 import io.littlehorse.common.exceptions.LHValidationError;
 import io.littlehorse.common.model.LHSerializable;
-import io.littlehorse.common.model.meta.SubNode;
-import io.littlehorse.common.model.meta.VariableAssignment;
+import io.littlehorse.common.model.meta.subnode.UserAssignment;
 import io.littlehorse.common.model.meta.usertasks.UTActionTrigger;
 import io.littlehorse.common.model.meta.usertasks.UserTaskDef;
 import io.littlehorse.common.model.wfrun.subnoderun.UserTaskNodeRun;
