+++ server/src/main/java/io/littlehorse/common/model/wfrun/UserTaskRun.java
@@ -10,8 +10,7 @@ import io.littlehorse.common.model.command.subcommand.AssignUserTaskRun;
 import io.littlehorse.common.model.command.subcommand.CompleteUserTaskRun;
 import io.littlehorse.common.model.command.subcommand.ReassignUserTask;
 import io.littlehorse.common.model.meta.Node;
-import io.littlehorse.common.model.meta.subnode.UserAssignment;
-import io.littlehorse.common.model.meta.subnode.UserTaskNode;
+import io.littlehorse.common.model.meta.UserTaskNode;
 import io.littlehorse.common.model.meta.usertasks.UTActionTrigger;
 import io.littlehorse.common.model.meta.usertasks.UserTaskDef;
 import io.littlehorse.common.model.objectId.NodeRunId;
